Moved NavigationData functions into StateController
