+++ NavigationJS/src/StateController.ts
@@ -5,7 +5,6 @@ import IDialog = require('./config/IDialog');
 import HashHistoryManager = require('./history/HashHistoryManager');
 import HistoryAction = require('./history/HistoryAction');
 import IHistoryManager = require('./history/IHistoryManager');
-import NavigationData = require('./NavigationData');
 import ReturnDataManager = require('./ReturnDataManager');
 import State = require('./config/State');
 import IState = require('./config/IState');
@@ -88,8 +87,8 @@ class StateController {
         if (this.stateContext.state) {
             this.stateContext.oldState = this.stateContext.state;
             this.stateContext.oldDialog = this.stateContext.dialog;
-            this.stateContext.oldData = NavigationData.clone(this.stateContext.data);
-            NavigationData.setDefaults(this.stateContext.oldData, this.stateContext.oldState.defaults);
+            this.stateContext.oldData = this.cloneData(this.stateContext.data);
+            this.setDataDefaults(this.stateContext.oldData, this.stateContext.oldState.defaults);
         }
     }
     
@@ -298,7 +297,7 @@ class StateController {
             if (key !== 'crumb' && !this.isDefault(key, data, state, !!separableData[key]))
                 newData[key] = ReturnDataManager.parseURLString(this.converterFactory, key, data[key], state, false, !!separableData[key]);
         }
-        NavigationData.setDefaults(newData, state.defaults);
+        this.setDataDefaults(newData, state.defaults);
         return newData;
     }
     
@@ -335,6 +334,20 @@ class StateController {
         if (distance > this.stateContext.crumbs.length || distance <= 0)
             throw new Error('The distance parameter must be greater than zero and less than or equal to the number of Crumbs (' + this.stateContext.crumbs.length + ')');
         return this.stateContext.crumbs[this.stateContext.crumbs.length - distance];
+    }
+    
+    private setDataDefaults(data: any, defaults: any) {
+        for (var key in defaults) {
+            if (data[key] == null || !data[key].toString())
+                data[key] = defaults[key];
+        }
+    }
+
+    private cloneData(data: any) {
+        var clone = {};
+        for (var key in data)
+            clone[key] = data[key];
+        return clone;
     }
     
     start(url?: string) {
